DS-2381: fix merge + shift queries into QueryBuilder.
